Make mobile app stay logged in on web
Closes #114
