+++ example/mobile_app/lib/app_state/app_state.dart
@@ -1,0 +1,1 @@
+export 'app_state_vm.dart' if (dart.library.js) 'app_state_web.dart';
