+++ example/mobile_app/lib/app_state/app_state_web.dart
@@ -1,0 +1,17 @@
+import 'dart:html';
+import 'package:flutter/foundation.dart';
+
+class AppState extends ChangeNotifier {
+  bool get isLoggedIn => window.sessionStorage['isLoggedIn'] == 'true';
+  set isLoggedIn(bool value) {
+    window.sessionStorage['isLoggedIn'] = value.toString();
+    notifyListeners();
+  }
+
+  bool _showBonusTab = false;
+  bool get showBonusTab => _showBonusTab;
+  set showBonusTab(bool value) {
+    _showBonusTab = value;
+    notifyListeners();
+  }
+}
