+++ example/mobile_app/lib/pages/home_page.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
-import 'package:mobile_app/main.dart';
+import 'package:mobile_app/app_state/app_state.dart';
 import 'package:provider/provider.dart';
 import 'package:routemaster/routemaster.dart';
