+++ example/mobile_app/lib/pages/login_page.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:mobile_app/main.dart';
+import 'package:mobile_app/app_state/app_state.dart';
 import 'package:provider/provider.dart';
 
 class LoginPage extends StatelessWidget {
