+++ example/mobile_app/lib/pages/tab_bar_page.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
-import 'package:mobile_app/main.dart';
+import 'package:mobile_app/app_state/app_state.dart';
 import 'package:provider/provider.dart';
 import 'package:routemaster/routemaster.dart';
